requirements.txt
matplotlib
numpy
opencv_python
pandas
PyYAML
scipy
Shapely
tensorboard
traci